+++ docs/CHANGES.rst
@@ -4,7 +4,13 @@ Changelog
 1.7.7 (unreleased)
 ------------------
 
-- TBD
+- Check the boolean value of the ``remove`` option when importing
+  objects.  Previously we only checked if the ``remove`` option was
+  given, regardless of its value.  Supported are ``True``, ``Yes``,
+  and ``1``, where case does not matter.  The syntax for removing
+  objects, properties, and elements is now the same.
+
+- Support ``remove="True"`` for properties.
 
 
 1.7.6 (2015-07-15)